feat: 상품 타입 정의 및 관련 파일 수정
- 상품 관련 타입을 index.ts 파일에 정의하여 코드의 일관성을 높였습니다.
- App.tsx, promotionService.ts, stockUtils.ts, ProductSelector.tsx에서 Product 타입을 사용하도록 수정하여 타입 안전성을 강화하였습니다.
- 불필요한 Product 인터페이스 정의를 제거하여 코드의 간결성을 향상시켰습니다.
